%methods doesn't need to list the number of %s...
Since %methods contains the string that is passed to sprintf(), you
do not need to pass the number of items needed. You only have to count
the number of %s in the string.

And I think you can use the goaste operator for that... ;-)

URL: https://rt.cpan.org/Ticket/Display.html?id=21298